HasUpdateConfiguration
