+++ db/src/Pos/DB/Block/Slog/Logic.hs
@@ -36,7 +36,7 @@ import           Pos.Chain.Block (Block, Blund, HasSlogGState, SlogUndo (..),
                      mainBlockSlot, prevBlockL, verifyBlocks)
 import           Pos.Chain.Genesis as Genesis (Config (..), configEpochSlots,
                      configK)
-import           Pos.Chain.Update (BlockVersion (..), HasUpdateConfiguration,
+import           Pos.Chain.Update (BlockVersion (..), UpdateConfiguration,
                      lastKnownBlockVersion)
 import           Pos.Core (BlockCount, FlatSlotId, ProtocolConstants,
                      difficultyL, epochIndexL, flattenSlotId, kEpochSlots,
@@ -58,10 +58,11 @@ import qualified Pos.DB.GState.Common as GS
                      getMaxSeenDifficulty)
 import           Pos.DB.Lrc (HasLrcContext, lrcActionOnEpochReason)
 import qualified Pos.DB.Lrc as LrcDB
-import           Pos.DB.Update (getAdoptedBVFull, getConsensusEra)
+import           Pos.DB.Update.GState (getAdoptedBVFull, getConsensusEra)
 import           Pos.Util (_neHead, _neLast)
 import           Pos.Util.AssertMode (inAssertMode)
 import           Pos.Util.Wlog (WithLogger, logInfo)
+import Pos.Util.Util (HasLens', lensOf')
 
 ----------------------------------------------------------------------------
 -- Helpers
@@ -91,12 +92,12 @@ import           Pos.Util.Wlog (WithLogger, logInfo)
 -- done only if `alt` component is the same as adopted one. In
 -- other cases (i. e. when our `(major, minor)` is less than from
 -- adopted version) check is not done.
-mustDataBeKnown :: HasUpdateConfiguration => BlockVersion -> Bool
-mustDataBeKnown adoptedBV =
-    lastKnownMajMin > adoptedMajMin || lastKnownBlockVersion == adoptedBV
+mustDataBeKnown :: UpdateConfiguration -> BlockVersion -> Bool
+mustDataBeKnown uc adoptedBV =
+    lastKnownMajMin > adoptedMajMin || lastKnownBlockVersion uc == adoptedBV
   where
     toMajMin BlockVersion {..} = (bvMajor, bvMinor)
-    lastKnownMajMin = toMajMin lastKnownBlockVersion
+    lastKnownMajMin = toMajMin (lastKnownBlockVersion uc)
     adoptedMajMin = toMajMin adoptedBV
 
 ----------------------------------------------------------------------------
@@ -109,7 +110,6 @@ type MonadSlogBase ctx m =
     , MonadIO m
     , MonadDBRead m
     , WithLogger m
-    , HasUpdateConfiguration
     )
 
 -- | Set of constraints needed for Slog verification.
@@ -117,6 +117,7 @@ type MonadSlogVerify ctx m =
     ( MonadSlogBase ctx m
     , MonadReader ctx m
     , HasLrcContext ctx
+    , HasLens' ctx UpdateConfiguration
     )
 
 -- | Verify everything from block that is not checked by other components.
@@ -138,7 +139,8 @@ slogVerifyBlocks genesisConfig curSlot blocks = runExceptT $ do
     era <- getConsensusEra
     logInfo $ sformat ("slogVerifyBlocks: Consensus era is " % shown) era
     (adoptedBV, adoptedBVD) <- lift getAdoptedBVFull
-    let dataMustBeKnown = mustDataBeKnown adoptedBV
+    uc <- view (lensOf' @UpdateConfiguration)
+    let dataMustBeKnown = mustDataBeKnown uc adoptedBV
     let headEpoch = blocks ^. _Wrapped . _neHead . epochIndexL
     leaders <- lift $
         lrcActionOnEpochReason
