+++ lib/src/Pos/WorkMode/Class.hs
@@ -16,12 +16,13 @@ import           Control.Monad.Trans.Control (MonadBaseControl)
 import qualified Crypto.Random as Rand
 import           UnliftIO (MonadUnliftIO)
 
-import           Pos.Chain.Block (HasBlockConfiguration, HasSlogContext,
-                     HasSlogGState, MonadLastKnownHeader)
+import           Pos.Chain.Block (HasSlogContext, HasSlogGState,
+                     MonadLastKnownHeader)
+import           Pos.Chain.Block (HasBlockConfiguration)
 import           Pos.Chain.Delegation (HasDlgConfiguration, MonadDelegation)
 import           Pos.Chain.Security (SecurityParams)
 import           Pos.Chain.Ssc (HasSscConfiguration, MonadSscMem)
-import           Pos.Chain.Update (HasUpdateConfiguration, UpdateParams)
+import           Pos.Chain.Update (UpdateConfiguration, UpdateParams)
 import           Pos.Configuration (HasNodeConfiguration)
 import           Pos.Context (BlockRetrievalQueue, BlockRetrievalQueueTag,
                      HasSscContext, StartTime, TxpGlobalSettings)
@@ -47,7 +48,7 @@ import           Pos.Util.Wlog (WithLogger)
 
 -- | Bunch of constraints to perform work for real world distributed system.
 type WorkMode ctx m
-    = ( MinWorkMode m
+    = ( MinWorkMode ctx m
       , MonadBaseControl IO m
       , Rand.MonadRandom m
       , MonadMask m
@@ -87,12 +88,13 @@ type WorkMode ctx m
       )
 
 -- | More relaxed version of 'WorkMode'.
-type MinWorkMode m
+type MinWorkMode ctx m
     = ( WithLogger m
       , CanJsonLog m
       , MonadIO m
       , MonadUnliftIO m
-      , HasUpdateConfiguration
+      , MonadReader ctx m
+      , HasLens' ctx UpdateConfiguration
       , HasNodeConfiguration
       , HasBlockConfiguration
       )
